+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
-## Unreleased changes
+## 3.0.0-beta.1
+* akashic-engine@3.0.0-beta.28 に追従
+
+## 3.0.0-beta.0
 * akashic-engine@3.0.0 系に追従。あわせてバージョンを 3.0.0 に。
 
 ## 2.3.0